# server/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /assets/ingredients/{filename}:
    get:
      operationId: StaticAssetsController_getIngredientImage
      parameters:
        - name: filename
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
...
  /ingredients/low-stock:
    get:
      operationId: IngredientsController_getLowStock
      parameters: []
      responses:
        '200':
          description: Returns list of ingredients below threshold
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
      security:
        - JWT-auth: []
      summary: Get ingredients with low stock
      tags: &ref_3
        - ingredients
  /ingredients:
    get:
      operationId: IngredientsController_findAll
      parameters:
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted ingredients (true/false)
          schema:
            type: string
      responses:
        '200':
          description: Returns list of all ingredients with current stock
      summary: Get all ingredients
      tags: *ref_3
    post:
      operationId: IngredientsController_create
      parameters: []
      requestBody:
        required: true
        description: Ingredient data with optional image
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Thịt bò
                unit:
                  type: string
                  example: kg
                threshold:
                  type: number
                  example: 10
                description:
                  type: string
                  example: Thịt bò tươi
                image:
                  type: string
                  format: binary
              required:
                - name
                - unit
                - threshold
      responses:
        '201':
          description: Ingredient created successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
      security:
        - JWT-auth: []
      summary: Create new ingredient
      tags: *ref_3
  /ingredients/{id}:
    get:
      operationId: IngredientsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Ingredient ID
          schema:
            type: string
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted ingredients (true/false)
          schema:
            type: string
      responses:
        '200':
          description: Returns ingredient details with current stock
        '404':
          description: Ingredient not found
      summary: Get ingredient by ID
      tags: *ref_3
    patch:
      operationId: IngredientsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Ingredient ID
          schema:
            type: string
      requestBody:
        required: true
        description: Updated ingredient data with optional image
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Thịt bò
                unit:
                  type: string
                  example: kg
                threshold:
                  type: number
                  example: 10
                description:
                  type: string
                  example: Thịt bò tươi
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Ingredient updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Ingredient not found
      security:
        - JWT-auth: []
      summary: Update ingredient
      tags: *ref_3
    delete:
      operationId: IngredientsController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Ingredient ID
          schema:
            type: string
      responses:
        '200':
          description: Ingredient deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Ingredient not found
      security:
        - JWT-auth: []
      summary: Delete ingredient (soft delete)
      tags: *ref_3
  /ingredients/{id}/restore:
    post:
      operationId: IngredientsController_restore
      parameters:
        - name: id
          required: true
          in: path
          description: Ingredient ID
          schema:
            type: string
      responses:
        '200':
          description: Ingredient restored successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Ingredient not found
      security:
        - JWT-auth: []
      summary: Restore deleted ingredient
      tags: *ref_3
...
info:
  title: Restaurant Management API
  description: |2-

        # Restaurant Management System API Documentation
        
        Hệ thống quản lý nhà hàng toàn diện với các tính năng đầy đủ cho việc vận hành nhà hàng hiện đại.
        
        ## Tính năng chính
        - **🔐 Xác thực & Phân quyền**: JWT-based authentication với phân quyền theo vai trò
        - **👥 Quản lý người dùng**: Admin, Staff, Kitchen, Waiter, Warehouse roles
        - **🍽️ Quản lý thực đơn**: Categories, dishes, menus với theo dõi nguyên liệu
        - **📋 Quản lý đơn hàng**: Đặt bàn qua QR, workflow bếp, theo dõi trạng thái
        - **📦 Quản lý kho**: Theo dõi nguyên liệu, quản lý lô hàng, nhập/xuất kho
        - **💳 Thanh toán**: Tích hợp VNPay cho thanh toán online
        - **🪑 Quản lý bàn**: Tạo QR code, theo dõi trạng thái bàn
        - **📊 Báo cáo**: Báo cáo tài chính, báo cáo kho
        
        ## Xác thực API
        Hầu hết các endpoint yêu cầu xác thực qua JWT token. Thêm token vào header Authorization:
        `Authorization: Bearer <your-jwt-token>`
        
        ## Phân quyền người dùng
        - **ADMIN**: Toàn quyền hệ thống
        - **STAFF**: Các thao tác nhân viên phục vụ
        - **KITCHEN**: Các thao tác đặc thù bếp
        - **WAITER**: Quản lý bàn và đơn hàng
        - **WAREHOUSE**: Quản lý kho
        
        ## HTTP Status Codes
        API trả về các mã trạng thái HTTP chuẩn:
        - **200**: Thành công
        - **201**: Tạo mới thành công
...
servers:
...
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Nhập JWT token (không cần prefix Bearer)
      in: header
...